requirements.txt
# Main application dependencies
openai==1.35.13
openai-whisper>=20231117
moviepy==1.0.3
ffmpeg-python==0.2.0
speechbrain==1.0.0

# Pexels API for B-Roll
pexels-api-py==0.0.5

# YouTube API & Auth
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# General utilities
python-dotenv==1.0.1
requests==2.32.3
httpx==0.26.0
numpy
tqdm
librosa

# PyTorch dependencies (CPU for now, adjust as needed)
torch>=2.6.0
torchaudio>=2.6.0
torchvision>=0.19.0

# Enhanced video editing features
sentence-transformers>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Added from the code block
auto-editor
imageio-ffmpeg
beautifulsoup4

# Babel
Babel

# Development & Linting
black==24.4.2
flake8==7.1.0
python-dotenv
